feat WelcomeDetailView, Debug environment variables
